+++ jdaviz/configs/cubeviz/plugins/spectral_extraction/spectral_extraction.vue
@@ -1,7 +1,7 @@
 <template>
   <j-tray-plugin
     :config="config"
-    :plugin_key="plugin_key || 'Spectral Extraction'"
+    :plugin_key="plugin_key || '3D Spectral Extraction'"
     :api_hints_enabled.sync="api_hints_enabled"
     :description="docs_description"
     :link="docs_link || 'https://jdaviz.readthedocs.io/en/'+vdocs+'/'+config+'/plugins.html#spectral-extraction'"
@@ -223,7 +223,7 @@
         />
         <v-row>
           <j-docs-link>
-              See the <j-external-link link='https://photutils.readthedocs.io/en/stable/aperture.html#aperture-and-pixel-overlap'
+              See the <j-external-link link='https://photutils.readthedocs.io/en/stable/user_guide/aperture.html#aperture-and-pixel-overlap'
               linktext='photutils docs'></j-external-link> for more details on aperture masking methods.
           </j-docs-link>
         </v-row>