+++ jdaviz/configs/specviz/plugins/unit_conversion/unit_conversion.py
@@ -304,7 +304,7 @@ class UnitConversion(PluginTemplateMixin):
             # NOTE: this assumes that all image data is coerced to surface brightness units
             layers = [lyr for lyr in msg.viewer.layers if lyr.layer.data.label == msg.data.label]
 
-            if not isinstance(data_obj, tracing.FlatTrace):
+            if not isinstance(data_obj, tracing.Trace):
 
                 if not len(self.spectral_unit_selected) and hasattr(data_obj, 'spectral_axis'):
                     try: