+++ jdaviz/core/loaders/test_loaders.py
@@ -4,6 +4,7 @@ from astropy import units as u
 from specutils import SpectralRegion, Spectrum
 from jdaviz.core.registries import loader_resolver_registry
 from jdaviz.core.loaders.resolvers import find_matching_resolver
+from jdaviz.utils import cached_uri
 
 
 def test_loaders_registry(specviz_helper):
@@ -52,7 +53,7 @@ def test_resolver_matching(specviz_helper):
 def test_trace_importer(specviz2d_helper, spectrum2d):
     specviz2d_helper._load(spectrum2d, format='2D Spectrum')
 
-    trace = specviz2d_helper.plugins['Spectral Extraction'].export_trace()
+    trace = specviz2d_helper.plugins['2D Spectral Extraction'].export_trace()
 
     res_sp = find_matching_resolver(specviz2d_helper.app, trace)
     assert res_sp._obj._registry_label == 'object'
@@ -81,8 +82,14 @@ def test_markers_specviz2d_unit_conversion(specviz2d_helper, spectrum2d):
 @pytest.mark.remote_data
 @pytest.mark.filterwarnings(r"ignore::astropy.wcs.wcs.FITSFixedWarning")
 def test_fits_spectrum2d(deconfigged_helper):
-    ldr = deconfigged_helper.loaders['url']
-    ldr.url = 'mast:jwst/product/jw02123-o001_v000000353_nirspec_f170lp-g235h_s2d.fits'
+    uri = cached_uri('mast:jwst/product/jw02123-o001_v000000353_nirspec_f170lp-g235h_s2d.fits')
+    if 'mast' in uri:
+        ldr = deconfigged_helper.loaders['url']
+        ldr.cache = True
+        ldr.url = uri
+    else:
+        ldr = deconfigged_helper.loaders['file']
+        ldr.filepath = uri
 
     # Default is Image but the test switches to 2D Spectrum
     # since this file type is not yet supported by the image loader
@@ -114,6 +121,12 @@ def test_resolver_url(deconfigged_helper):
     loader.url = 'not-valid-url'
     assert len(loader.format.choices) == 0
 
+    # s3 input
+    loader.url = "s3://stpubdata/jwst/public/jw02727/L3/t/o002/jw02727-o002_t062_nircam_clear-f277w_i2d.fits"  # noqa: E501
+    assert loader._obj.url_scheme == 's3'
+    assert len(loader.format.choices) > 0
+
+    # https valid input
     loader.url = 'https://stsci.box.com/shared/static/exnkul627fcuhy5akf2gswytud5tazmw.fits'  # noqa
     assert len(loader.format.choices) == 4  # may change with future importers
     assert loader.format.selected == 'Image'  # default may change with future importers
@@ -141,6 +154,8 @@ def test_resolver_url(deconfigged_helper):
     assert len(deconfigged_helper.viewers) == 2
 
     with pytest.raises(ValueError, match="Failed query for URI"):
+        # NOTE: this test will attempt to reach out to MAST via astroquery
+        # even if cache is available.
         deconfigged_helper.load('mast:invalid')